requirements.txt
Flask==2.0.2
#pytest==7.0.1; python_version > "3.0"
# pin pytest to 4.6.11 for Python2.
#pytest==7.0.1; python_version < "3.0"
gunicorn==20.1.0
google-cloud-pubsub==2.11.0
google-cloud-storage==2.2.1
pysftp==0.2.9